+++ CMakeLists.txt
@@ -30,7 +30,7 @@ add_executable(basic_rt3 ${RT3_SOURCE_DIR}/core/api.cpp
                          ${RT3_SOURCE_DIR}/core/image_io.cpp
                          ${RT3_SOURCE_DIR}/core/parser.cpp
                          ${RT3_SOURCE_DIR}/cameras/camera.cpp
-                         ${RT3_SOURCE_DIR}/core/integrator.cpp
+                         ${RT3_SOURCE_DIR}/integrators/integrator.cpp
                          ${RT3_SOURCE_DIR}/core/material.cpp
                          ${RT3_SOURCE_DIR}/core/ray.cpp
                          ${RT3_SOURCE_DIR}/core/sphere.cpp
