+++ src/core/rt3-base.h
@@ -3,9 +3,9 @@
 
 #include "../ext/lodepng.h"
 #include "error.h"
-#include "film.h"
+#include "../cameras/film.h"
 #include "background.h"
-#include "camera.h"
+#include "../cameras/camera.h"
 #include "image_io.h"
 #include "paramset.h"
 #include "parser.h"
@@ -16,6 +16,6 @@
 #include "shape.h"
 #include "primlist.h"
 #include "scene.h"
-#include "integrator.h"
+#include "../integrators/integrator.h"
 
 #endif  // RT3-BASE_H
