+++ src/integrators/integrator.h
@@ -1,13 +1,13 @@
 #ifndef INTEGRATOR_H
 #define INTEGRATOR_H 1
 
-#include "primitive.h"
-#include "ray.h"
-#include "rt3.h"
-#include "scene.h"
+#include "../core/primitive.h"
+#include "../core/ray.h"
+#include "../core/rt3.h"
+#include "../core/scene.h"
 #include "../cameras/camera.h"
-#include "background.h"
-#include "material.h"
+#include "../core/background.h"
+#include "../core/material.h"
 #include <optional>
 
 namespace rt3{
